- Assimp libs for debug mode added - All warnings are removed
Former-commit-id: 16da9edb6363afec7b3d67870e57d862777c5570 [formerly c47bf35970e28440400145ea991cf62ad051be15]
Former-commit-id: c39c47402044f9eaa05d1cbddc069d04db552871

+++ Renderer/Renderer/Model.cpp
@@ -135,7 +135,7 @@ std::vector<Texture> Model::LoadMaterialTextures(aiMaterial * mat, aiTextureType
         mat->GetTexture(type, i, &str);
 
         bool skip = false;
-        unsigned int texture_loaded_size = textures_loaded.size();
+        unsigned int texture_loaded_size = static_cast<unsigned int>(textures_loaded.size());
         for (unsigned int j = 0; j < texture_loaded_size; ++j)
         {
             if (textures_loaded[j].path == std::string(str.C_Str()))
